+++ src/main/java/bowling/exception/BowlingException.java
@@ -1,6 +1,14 @@
 package bowling.exception;
 
-public class BowlingException extends RuntimeException{
+// TODO(jack.comeback) : code 추가
+public class BowlingException extends RuntimeException {
+
+    public BowlingException() {
+    }
+
+    public BowlingException(String message) {
+        super(message);
+    }
 
     public BowlingException(BowlingExceptionCode exceptionCode, String value) {
         super(String.format("%s value - %s", exceptionCode.getMessage(), value));
