+++ src/main/java/bowling/presentation/Format.java
@@ -1,6 +1,6 @@
 package bowling.presentation;
 
-import bowling.domain.frame.FallenPins;
+import java.util.List;
 
 import static bowling.presentation.PinFormat.*;
 import static java.math.BigInteger.TEN;
@@ -8,7 +8,7 @@ import static java.math.BigInteger.ZERO;
 
 // board에 출력되는 format을 만드는 객체
 public class Format {
-    public static String fallenPinsFormat(FallenPins fallenPins) {
+    public static String fallenPinsFormat(List<Integer> fallenPins) {
         if (fallenPins.size() == 1) {
             return singlePinInBoard(fallenPins);
         }
@@ -18,25 +18,29 @@ public class Format {
         if (fallenPins.size() == 3) {
             return triplePinInBoard(fallenPins);
         }
+        return emptyFormat();
+    }
+
+    public static String emptyFormat() {
         return String.format(centerAlignFormat(6), "");
     }
 
-    private static String singlePinInBoard(FallenPins fallenPins) {
-        return String.format(centerAlignFormat(3), singlePinFormat(fallenPins.first()));
+    private static String singlePinInBoard(List<Integer> fallenPins) {
+        return String.format(centerAlignFormat(3), singlePinFormat(fallenPins.get(0)));
     }
 
-    private static String singlePinFormat(int score) {
-        if (TEN.intValue() == score) {
+    private static String singlePinFormat(int pin) {
+        if (TEN.intValue() == pin) {
             return STRIKE.format();
         }
-        if (ZERO.intValue() == score) {
+        if (ZERO.intValue() == pin) {
             return GUTTER.format();
         }
-        return String.valueOf(score);
+        return String.valueOf(pin);
     }
 
-    private static String doublePinInBoard(FallenPins fallenPins) {
-        return String.format(centerAlignFormat(4), doublePinFormat(fallenPins.first(), fallenPins.second()));
+    private static String doublePinInBoard(List<Integer> fallenPins) {
+        return String.format(centerAlignFormat(4), doublePinFormat(fallenPins.get(0), fallenPins.get(1)));
     }
 
     private static String doublePinFormat(int first, int second) {
@@ -46,8 +50,8 @@ public class Format {
         return singlePinFormat(first) + DELIMITER.format() + singlePinFormat(second);
     }
 
-    private static String triplePinInBoard(FallenPins fallenPins) {
-        return String.format(centerAlignFormat(5), triplePinFormat(fallenPins.first(), fallenPins.second(), fallenPins.third()));
+    private static String triplePinInBoard(List<Integer> fallenPins) {
+        return String.format(centerAlignFormat(5), triplePinFormat(fallenPins.get(0), fallenPins.get(1), fallenPins.get(2)));
     }
 
     private static String triplePinFormat(int first, int second, int third) {
@@ -72,11 +76,21 @@ public class Format {
         return DELIMITER.format() + String.format(centerAlignFormat(4), name) + DELIMITER.format();
     }
 
+    public static String scoreName(String name) {
+        return DELIMITER.format() + String.format(centerAlignFormat(5), name) + DELIMITER.format();
+    }
+
     private static String centerAlignFormat(int size) {
         int MAX_WIDTH = 6;
         if (size > MAX_WIDTH) {
             throw new IllegalArgumentException(String.format("size는 %d보다 작아야 합니다. - %d", MAX_WIDTH, size));
         }
         return "%" + size + "s" + " ".repeat(MAX_WIDTH - size);
+    }
+
+    public static String scoreFormat(int score) {
+        return score == 0
+                ? String.format(centerAlignFormat(6), "")
+                : String.format(centerAlignFormat(3), String.format("%3d", score));
     }
 }
