--- src/main/java/bowling2/presentation/PinFormat.java
@@ -1,21 +1,0 @@
-package bowling2.presentation;
-
-public enum PinFormat {
-    // score
-    STRIKE("X"),
-    SPARE("/"),
-    GUTTER("-"),
-
-    // line
-    DELIMITER("|");
-
-    private final String format;
-
-    PinFormat(String format) {
-        this.format = format;
-    }
-
-    public String format() {
-        return format;
-    }
-}
