+++ src/test/java/bowling/domain/ScorePendingQueueTest.java
@@ -1,4 +1,4 @@
-package bowling2.domain;
+package bowling.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
