+++ src/test/java/bowling/domain/score/CommonScoreStrategyTest.java
@@ -1,7 +1,7 @@
-package bowling2.domain.score;
+package bowling.domain.score;
 
-import bowling2.domain.frame.FinalFrame;
-import bowling2.domain.frame.NormalFrame;
+import bowling.domain.frame.FinalFrame;
+import bowling.domain.frame.NormalFrame;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
